fix(dropdown): fix misplaced native dropdown label
Labels of native dropdowns were being placed with position absolute relative to the nearest
positioned ancestor of the dropdown instead of the dropdown itself, since the dropdown class had
position set to static. This commit sets positon relative on the dropdown, so the label is
positioned relative to it.

AB#1826