fix(header): stop dropdowns showing briefly on page load (#484)
This is a partial quick fix to a larger issue. The issue seems to happen on anything with a
transition property defined on it. If i add a background color transition on sdds-btn, for example,
it will fade in from the browser defaults on page load. I think its because the css is loaded in
sdds-theme, so it might come in a bit late.

re #347, AD#1931