fix(button): corrected the styling of secondary, ghost and danger button (#482)
* fix(button): corrected the styling of seconday,ghost and danger button

The focus style was wrong for all of the buttons above. The border should be 2px according to
design, however it was 1px. The secondary button also had the wrong background, border-color and
color for hover-state.

* fix(button): used the mixin for outline instead of recalculating the height and padding

used the mixin for outline instead of recalculating the height and padding. Also added the variable
--sdds-btn-primary-outline-color-focus to use the correct outline color on all the buttons.