fix(date-time): added props for label/helper and removed slots, also introdues default-value
Removed the slots for label/helper and created props for them instead, also introduces the prop
default-value to allow the user to select a default time to be displayed.

BREAKING CHANGE: Removed the slots for label/helper and created props for them instead, also
introduces the prop default-value to allow the user to select a default time to be displayed.